Fix AppBasicaCRUD sample compile
