+++ AgentWorker/output/appbasicacrud/Layout/MainLayout.razor
@@ -1,5 +1,6 @@
+@namespace AppBasicaCRUD.Layout
 @inherits LayoutComponentBase
-@using Appbasicacrud.Layout // Ensure using for NavMenu
+@using AppBasicaCRUD.Layout
 
 <div class="page">
     <div class="sidebar">
