+++ AgentWorker/output/appbasicacrud/Layout/NavMenu.razor
@@ -1,7 +1,9 @@
 
+@namespace AppBasicaCRUD.Layout
+
 <div class="top-row ps-3 navbar navbar-dark">
     <div class="container-fluid">
-        <a class="navbar-brand" href="">Appbasicacrud</a>
+        <a class="navbar-brand" href="">AppBasicaCRUD</a>
     </div>
 </div>
 
@@ -15,7 +17,12 @@
             </NavLink>
         </div>
 
-        @* Add other navigation links here *@
+        <div class="nav-item px-3">
+            <NavLink class="nav-link" href="/productos">
+                <span class="bi bi-list" aria-hidden="true"></span> Productos
+            </NavLink>
+        </div>
 
     </nav>
 </div>
+
