+++ AgentWorker/output/appbasicacrud/Pages/Create.razor
@@ -1,6 +1,8 @@
 @page "/create"
 @inject ApplicationDbContext dbContext
 @inject NavigationManager NavigationManager
+@using AppBasicaCRUD.Models
+@using AppBasicaCRUD.Data
 
 <h3>Crear Producto</h3>
