+++ AgentWorker/output/appbasicacrud/Pages/Edit.razor
@@ -1,20 +1,22 @@
 @page "/edit/{id:int}"
 @inject ApplicationDbContext dbContext
 @inject NavigationManager NavigationManager
+@using AppBasicaCRUD.Models
+@using AppBasicaCRUD.Data
 
 <h3>Edit</h3>
 
-<EditForm Model="@person" OnValidSubmit="HandleValidSubmit">
+<EditForm Model="@producto" OnValidSubmit="HandleValidSubmit">
     <DataAnnotationsValidator />
     <ValidationSummary />
 
     <div class="form-group">
-        <label for="Name">Name</label>
-        <InputText id="Name" class="form-control" @bind-Value="person.Name" />
+        <label for="Nombre">Nombre</label>
+        <InputText id="Nombre" class="form-control" @bind-Value="producto.Nombre" />
     </div>
     <div class="form-group">
-        <label for="Age">Age</label>
-        <InputNumber id="Age" class="form-control" @bind-Value="person.Age" />
+        <label for="Precio">Precio</label>
+        <InputNumber id="Precio" class="form-control" @bind-Value="producto.Precio" />
     </div>
     <button type="submit" class="btn btn-primary">Submit</button>
 </EditForm>
@@ -23,18 +25,18 @@
     [Parameter]
     public int id { get; set; }
 
-    Person person = new Person();
+    Producto producto = new Producto();
 
     protected override async Task OnInitializedAsync()
     {
-        person = await dbContext.People.FindAsync(id);
+        producto = await dbContext.Productos.FindAsync(id);
     }
 
     private async void HandleValidSubmit()
     {
-        dbContext.People.Update(person);
+        dbContext.Productos.Update(producto);
         await dbContext.SaveChangesAsync();
 
-        NavigationManager.NavigateTo("/");
+        NavigationManager.NavigateTo("/productos");
     }
 }
