+++ AgentWorker/output/appbasicacrud/Routes.razor
@@ -1,4 +1,4 @@
-@using AppBasicaCRUD.Components
+@using AppBasicaCRUD.Layout
 <Router AppAssembly="@typeof(Program).Assembly">
     <Found Context="routeData">
         <RouteView RouteData="@routeData" DefaultLayout="@typeof(MainLayout)" />
