+++ AgentWorker/output/appbasicacrud/_Imports.razor
@@ -4,8 +4,6 @@
 @using Microsoft.AspNetCore.Components.Routing
 @using Microsoft.AspNetCore.Components.Web
 @using Microsoft.AspNetCore.Components.Web.Virtualization
-@using Microsoft.AspNetCore.Components.WebAssembly.Http
 @using Microsoft.JSInterop
 @using AppBasicaCRUD
-@using AppBasicaCRUD.Shared
-@using AppBasicaCRUD.Components
+@using AppBasicaCRUD.Models
